api/requirements.txt
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-DotEnv==0.1.2
google-api-core==1.17.0
google-auth==1.14.0
google-cloud-secret-manager==0.2.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.7
Werkzeug==0.16.0
